Improve trait split::Take
Renamed method take to take_range to avoid ambiguity.

Added method remove_range that can be overridden to optimize
removal of leading or trailing parts of the buffer without having to
return the split-off part.

Turned the generic parameter into associated type Slice. It would be
ergonomicallly bad to have two different Take impls for the same
type, and a crate that adds ambiguous impls spoils it for any crate
down the dependency graph.